+++ wgpu-types/src/lib.rs
@@ -6871,6 +6871,9 @@ pub enum ExternalImageSource {
     ///
     /// Requires [`DownlevelFlags::UNRESTRICTED_EXTERNAL_TEXTURE_COPIES`]
     OffscreenCanvas(web_sys::OffscreenCanvas),
+    /// Copy from a video frame.
+    #[cfg(web_sys_unstable_apis)]
+    VideoFrame(web_sys::VideoFrame),
 }
 
 #[cfg(target_arch = "wasm32")]
@@ -6884,6 +6887,8 @@ impl ExternalImageSource {
             ExternalImageSource::ImageData(i) => i.width(),
             ExternalImageSource::HTMLCanvasElement(c) => c.width(),
             ExternalImageSource::OffscreenCanvas(c) => c.width(),
+            #[cfg(web_sys_unstable_apis)]
+            ExternalImageSource::VideoFrame(v) => v.display_width(),
         }
     }
 
@@ -6896,6 +6901,8 @@ impl ExternalImageSource {
             ExternalImageSource::ImageData(i) => i.height(),
             ExternalImageSource::HTMLCanvasElement(c) => c.height(),
             ExternalImageSource::OffscreenCanvas(c) => c.height(),
+            #[cfg(web_sys_unstable_apis)]
+            ExternalImageSource::VideoFrame(v) => v.display_height(),
         }
     }
 }
@@ -6912,6 +6919,8 @@ impl std::ops::Deref for ExternalImageSource {
             Self::ImageData(i) => i,
             Self::HTMLCanvasElement(c) => c,
             Self::OffscreenCanvas(c) => c,
+            #[cfg(web_sys_unstable_apis)]
+            Self::VideoFrame(v) => v,
         }
     }
 }